Added SAINT to the model suite. Need to test components because this trully is a monster model
