+++ pytorch_widedeep/models/__init__.py
@@ -5,4 +5,7 @@ from pytorch_widedeep.models.wide_deep import WideDeep
 from pytorch_widedeep.models.deep_image import DeepImage
 from pytorch_widedeep.models.tab_resnet import TabResnet
 from pytorch_widedeep.models.tabnet.tab_net import TabNet
-from pytorch_widedeep.models.tab_transformer import TabTransformer
+from pytorch_widedeep.models.transformers.transformer_models import (
+    SAINT,
+    TabTransformer,
+)
